[#210] Move node builder FFI functions to own module
